[#74][Security] - set token validity to 30 minutes - removed token validation on logout route, as this route only unsets jwt cookie tokens only (should not be a security concern if accessed without authorization) - refactored code for token verification and generation into tokenRepository.js
