+++ src/repository/tokenRepository.js
@@ -1,0 +1,38 @@
+const jwt = require('jsonwebtoken');
+
+const verifyJWTAuth = (request, response, next) => {
+  const headerToken = request.headers.cookie;
+
+  if (headerToken == null) {
+    // No token: 403
+    response.status(403).json({ auth: false, message: 'No token found in request' });
+  } else {
+    jwt.verify(headerToken.split("=")[1], process.env.ACCESS_TOKEN_SECRET, {}, (err, decoded) => {
+      if (err) {
+          // Bad token: 403
+          response.status(401).json({ auth: false, message: 'User authentication failed' });
+      } else {
+        request.userId = decoded.id;
+
+        next();
+      }
+    });
+  }
+};
+
+const generateToken = (response, userId) => {
+  const tokenPayload = {
+    id: userId,
+  };
+
+  // generating the JWT token
+  const tokenJWT = jwt.sign(tokenPayload, process.env.ACCESS_TOKEN_SECRET, { expiresIn: '15s' }, {});
+
+  response.cookie('jwt', tokenJWT, {
+    httpOnly: true,
+    secure: true,
+    sameSite: "lax"
+  })
+}
+
+module.exports = { verifyJWTAuth, generateToken };
